pyproject.toml
[tool.poetry]
name = "restudy-drf-tutorial"
version = "0.1.0"
description = ""
authors = ["parkhongbeen <pack122@naver.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "<3"
djangorestframework = "^3.11.0"
notebook = "^6.0.3"
django-extensions = "^2.2.8"
model_bakery = "^1.1.0"
pygments = "^2.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
